requirements.txt
werkzeug==0.16.0
pillow==6.2.1
pytesseract==0.3.0
opencv-python==4.1.1.26
flask==1.1.1
gunicorn==19.9.0
python-dateutil==2.8.1
python-dotenv==0.10.3
sentry-sdk[flask]==0.13.1
jsonschema
itsdangerous==1.1.0
flask-marshmallow==0.10.1
marshmallow==3.2.1
requests==2.22.0
numpy==1.19.1